src/bot/requirements.txt
aiocache==0.12.3
APScheduler==3.11.0
ballpark @ git+https://github.com/Mee6/python-ballpark.git
chat-exporter==2.8.4
coc.py @ git+https://github.com/mathsman5133/coc.py.git@march-update
disnake @ git+https://github.com/MagicTheDev/disnake.git
disnake-ext-fluent @ git+https://github.com/MagicTheDev/disnake-ext-fluent.git
dnspython==2.7.0
docker==7.1.0
emoji==2.14.1
expiring-dict==1.1.2
fastapi==0.116.1
fluent.runtime==0.4.0
hashids==1.3.1
imagehash==4.3.2
ipython==9.4.0
kaleido==1.0.0
loguru==0.7.3
matplotlib==3.10.5
meilisearch-python-sdk==4.8.0
motor==3.7.1
msgspec==0.19.0
numerize==0.12
numpy==2.3.2
openai==1.100.2
openpyxl==3.1.5
orjson==3.11.2
pandas==2.3.1
pendulum==3.1.0
plotly==6.3.0
pymitter==1.1.3
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-snappy
pytz==2025.2
redis==6.4.0
sentry_sdk==2.35.0
stringcase==1.2.0
ujson
uvicorn==0.35.0
Wand==0.6.13
websockets==15.0.1
XlsxWriter==3.2.5
aiohttp==3.10.5
sphinx==8.1.3
pytest==8.3.2
prometheus-client==0.21.0
